creation of go-hardware/ghw library
Modified from jaypipes/ghw library in a number of ways:

First, we overhaul the way context is passed around and how options are
processed. No more weird setup context for snapshotting. The new
`ghw.NewContext()` method and associated `ghw.WithXXX()` functions are
all that is needed now. This means that all module functions (e.g.
`ghw.CPU()` or `ghw.Memory()`) accept a single parameter of type
`context.Context` and not some custom context-like pointer.

Second, we overhaul and dramatically simplify the snapshotting code. We
remove the `pkg/snapshot` and put all the creation and extraction code
into `cmd/ghw/snapshot`. There are also only two top-level
snapshot-related functions now: `snapshot.New()` and
`snapshot.Expand()`. The caller is now expected to handle the creation
and teardown of target expansion destination directories themselves.
This allowed us to get rid of the wonky dual-context stuff and simplify
the way that options and paths were constructed.

Snapshotting is now done by the `ghw snapshot` command instead of a
separate tool. Likewise, to inspect a snapshot, the `ghw -s <SNAPSHOT>`
flag accepts a path to a snapshot archive and informs the `ghw` CLI tool
to inspect the snapshot not the host system.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>